transaction supabase service

+++ flutter/lib/service/supabase/category_amount_supabase.dart
@@ -15,8 +15,8 @@ import '../category.dart';
 import '../category_amount.dart';
 import '../storage.dart';
 import '../validator.dart';
-import 'supabase.dart';
 import 'category_supabase.dart';
+import 'supabase.dart';
 
 const categoryAmountTable = 'categories_amounts';
 const lastUsedPeriodKey = 'last_used_period';
@@ -34,38 +34,36 @@ class CategoryAmountSupabaseService implements CategoryAmountService {
 
   @override
   Future<CategoryAmount> saveAmount({
-    required String categoryCode,
+    required Category category,
     String? amountCode,
     required Period period,
     required double amount,
   }) async {
     _saveLastUsed(period);
 
-    final user = _fetchUser();
-    final category = await _fetchCategoryByCode(categoryCode);
+    if (category is! SupabaseCategory) {
+      throw ValidationError({
+        'category': CategoryError.invalidCategory,
+      });
+    }
 
-    final categoryAmount =
-        _CategoryAmount(category: category, period: period, amount: amount);
+    final user = DI().get<AuthService>().fetchUser(errorIfMissing: CategoryError.invalidUser);
+
+    final categoryAmount = _CategoryAmount(category: category, period: period, amount: amount);
     final errors = amountValidator?.validate(categoryAmount);
     if (errors?.isNotEmpty ?? false) {
       throw ValidationError(errors!);
     }
 
-    final categoryAmountExists =
-        await _categoryAmountExistsByCategoryAndPeriod(category, period);
+    final categoryAmountExists = await _categoryAmountExistsByCategoryAndPeriod(category, period);
     if (categoryAmountExists) {
-      await config.supabase
-          .from(categoryAmountTable)
-          .update(categoryAmount.toMap(user))
-          .match({
+      await config.supabase.from(categoryAmountTable).update(categoryAmount.toMap(user)).match({
         categoryIdField: category.id,
         fromDateField: period.from.toIso8601String(),
         toDateField: period.to.toIso8601String(),
       });
     } else {
-      await config.supabase
-          .from(categoryAmountTable)
-          .insert(categoryAmount.toMap(user));
+      await config.supabase.from(categoryAmountTable).insert(categoryAmount.toMap(user));
     }
     return categoryAmount;
   }
@@ -97,7 +95,7 @@ class CategoryAmountSupabaseService implements CategoryAmountService {
 
     if (data is List) {
       final futureList = data.map((item) {
-        return _CategoryAmount.from(item, _fetchCategoryById);
+        return _CategoryAmount.from(item, fetcher: _fetchCategoryById);
       });
       final list = await Future.wait(futureList);
       return list.whereType<CategoryAmount>().toList();
@@ -107,12 +105,16 @@ class CategoryAmountSupabaseService implements CategoryAmountService {
 
   @override
   Future<void> deleteAmount({
-    required String categoryCode,
+    required Category category,
     required Period period,
   }) async {
     _saveLastUsed(period);
 
-    final category = await _fetchCategoryByCode(categoryCode);
+    if (category is! SupabaseCategory) {
+      throw ValidationError({
+        'category': CategoryError.invalidCategory,
+      });
+    }
     await config.supabase.from(categoryAmountTable).delete().match({
       categoryIdField: category.id,
       fromDateField: period.from.toIso8601String(),
@@ -177,8 +179,7 @@ class CategoryAmountSupabaseService implements CategoryAmountService {
     storageService.writeString(lastUsedPeriodKey, period.toString());
   }
 
-  Future<bool> _categoryAmountExistsByCategoryAndPeriod(
-      SupabaseCategory category, Period period) async {
+  Future<bool> _categoryAmountExistsByCategoryAndPeriod(SupabaseCategory category, Period period) async {
     final count = await config.supabase
         .from(categoryAmountTable)
         .select(
@@ -191,27 +192,6 @@ class CategoryAmountSupabaseService implements CategoryAmountService {
         .eq(fromDateField, period.from.toIso8601String())
         .eq(toDateField, period.to.toIso8601String());
     return count.count != null && count.count! > 0;
-  }
-
-  AppUser _fetchUser() {
-    final user = DI().get<AuthService>().user();
-    if (user != null) {
-      return user;
-    }
-    throw ValidationError<CategoryError>({
-      'user': CategoryError.invalidUser,
-    });
-  }
-
-  Future<SupabaseCategory> _fetchCategoryByCode(String code) async {
-    final category =
-        await DI().get<CategoryService>().fetchCategoryByCode(code);
-    if (category is SupabaseCategory) {
-      return category;
-    }
-    throw ValidationError({
-      'category': CategoryError.invalidCategory,
-    });
   }
 
   Future<SupabaseCategory?> _fetchCategoryById(int id) async {
@@ -256,26 +236,20 @@ class _CategoryAmount implements CategoryAmount {
   }
 
   static Future<_CategoryAmount?> from(
-      dynamic raw, TypedFutureFetcher<SupabaseCategory, int> fetcher) async {
+    dynamic raw, {
+    required TypedFutureFetcher<SupabaseCategory, int> fetcher,
+  }) async {
     if (raw is Map<String, dynamic>) {
       final id = raw[idField] as int?;
       final categoryId = raw[categoryIdField] as int?;
       final fromDate = DateTime.tryParse((raw[fromDateField] as String?) ?? '');
       final toDate = DateTime.tryParse((raw[toDateField] as String?) ?? '');
       final amount = raw[amountField] as num?;
-      if (id != null &&
-          categoryId != null &&
-          fromDate != null &&
-          toDate != null &&
-          amount != null) {
+      if (id != null && categoryId != null && fromDate != null && toDate != null && amount != null) {
         final category = await fetcher(categoryId);
         if (category != null) {
           final period = Period(from: fromDate, to: toDate);
-          return _CategoryAmount(
-              id: id,
-              category: category,
-              period: period,
-              amount: amount.toDouble());
+          return _CategoryAmount(id: id, category: category, period: period, amount: amount.toDouble());
         }
       }
     }
@@ -287,9 +261,7 @@ class _CategoryAmount implements CategoryAmount {
     if (identical(this, other)) {
       return true;
     }
-    return other is _CategoryAmount &&
-        runtimeType == other.runtimeType &&
-        category == other.category;
+    return other is _CategoryAmount && runtimeType == other.runtimeType && category == other.category;
   }
 
   @override
@@ -299,7 +271,6 @@ class _CategoryAmount implements CategoryAmount {
 
   @override
   CategoryAmount copyWith({required Period period}) {
-    return _CategoryAmount(
-        id: id, category: _category, period: period, amount: amount);
+    return _CategoryAmount(id: id, category: _category, period: period, amount: amount);
   }
 }
